+++ src/r0/push/get_pushrules_all.rs
@@ -1,10 +1,7 @@
 //! [GET /_matrix/client/r0/pushrules/](https://matrix.org/docs/spec/client_server/r0.6.0#get-matrix-client-r0-pushrules)
 
-use std::collections::BTreeMap;
-
 use ruma_api::ruma_api;
-
-use super::{PushRule, RuleKind};
+use ruma_events::push_rules::Ruleset;
 
 ruma_api! {
     metadata {
@@ -20,7 +17,7 @@ ruma_api! {
 
     response {
         /// The global ruleset
-        pub global: BTreeMap<RuleKind, Vec<PushRule>>
+        pub global: Ruleset,
     }
 
     error: crate::Error